backend/app/schemas/user.py
from __future__ import annotations

from datetime import datetime
from uuid import UUID

from pydantic import BaseModel, EmailStr, Field

from app.schemas.common import ORMModel


class UserRead(ORMModel):
    id: UUID
    username: str
    full_name: str | None
    email: str | None
    is_active: bool
    is_superuser: bool
    auth_source: str
    created_at: datetime


class UserCreate(BaseModel):
    username: str = Field(..., min_length=3, max_length=128)
    password: str = Field(..., min_length=8, max_length=128)
    email: EmailStr | None = None
    full_name: str | None = Field(default=None, max_length=255)
    is_superuser: bool = False
