backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pyodbc==5.2.0
redis==6.4.0
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
PyJWT[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
prometheus-client==0.22.1
pytest==8.4.2
